resource/database.py
import os

from dotenv import load_dotenv
from sqlmodel import Session, SQLModel, create_engine

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL",
                         "postgresql://postgres:91a07c05a48bb9ac7873@uvxtdw.easypanel.host:5430/belalady")

engine = create_engine(DATABASE_URL, echo=True)

def get_session():
    with Session(engine) as session:
        yield session
